+++ pkg/operator/apis/monitoring/v1alpha1/types.go
@@ -102,12 +102,12 @@ type GrafanaAgentSpec struct {
 	ServiceAccountName string `json:"serviceAccountName,omitempty"`
 	// Secrets is a list of secrets in the same namespace as the GrafanaAgent
 	// object which will be mounted into each running Grafana Agent pod.
-	// The secrets are mounted into /etc/grafana-agent/secrets/<secret-name>.
+	// The secrets are mounted into /etc/grafana-agent/extra-secrets/<secret-name>.
 	Secrets []string `json:"secrets,omitempty"`
 	// ConfigMaps is a liset of config maps in the same namespace as the
 	// GrafanaAgent object which will be mounted into each running Grafana Agent
 	// pod.
-	// The ConfigMaps are mounted into /etc/grafana-agent/configmaps/<configmap-name>.
+	// The ConfigMaps are mounted into /etc/grafana-agent/extra-configmaps/<configmap-name>.
 	ConfigMaps []string `json:"configMaps,omitempty"`
 	// Affinity, if specified, controls pod scheduling constraints.
 	Affinity *v1.Affinity `json:"affinity,omitempty"`
@@ -156,6 +156,8 @@ type GrafanaAgentSpec struct {
 	// +kubebuilder:default=false
 	EnableConfigReadAPI bool `json:"enableConfigReadAPI,omitempty"`
 }
+
+// +kubebuilder:object:generate=false
 
 // ObjectSelector is a set of selectors to use for finding an object in the
 // resource hierarchy. When NamespaceSelector is nil, objects should be